move util.js to lib/bbox.js
