+++ lib/bbox.js
